view orders list + order details

+++ lib/features/cart/screens/cart_screen.dart
@@ -22,7 +22,7 @@ class _CartScreenState extends State<CartScreen> {
         arguments: searchQuery);
   }
 
-  void naviagteToAddressScreen(int sum) {
+  void naviagteToAddressScreen(double sum) {
     Navigator.pushNamed(
       context,
       AddressScreen.routeName,
@@ -33,9 +33,9 @@ class _CartScreenState extends State<CartScreen> {
   @override
   Widget build(BuildContext context) {
     final user = context.watch<UserProvider>().user;
-    int sum = 0;
+    double sum = 0;
     user.cart
-        .map((e) => sum += e['quantity'] * e['product']['price'] as int)
+        .map((e) => sum += e['quantity'] * e['product']['price'] as double)
         .toList();
     return Scaffold(
       appBar: PreferredSize(
