+++ lib/features/cart/widgets/cart_subtotal.dart
@@ -8,10 +8,10 @@ class CartSubtotal extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     final user = context.watch<UserProvider>().user;
-    int sum = 0;
+    double sum = 0;
     user.cart
         .map(
-          (e) => sum += e['quantity'] * e['product']['price'] as int,
+          (e) => sum += e['quantity'] * e['product']['price'] as double,
         )
         .toList();
     return Container(
